Reworked schematic to include an ESP32-S2-MINI-2.
Reworked schematic to include an ESP32-S2-MINI-2 as the main CPU.

Two track outputs are directly on this board.  The rest of the GPIO
pins are routed so that the existing EX-MotorShield8874 can stack.
The result is potentially 6 power outputs from a single unit.

Added a USB interface to the module.

Moved the drivers into heirarchal sheets.

Removed all unnecessary components.

Added a JTAG header for debugging.